added rout to get all zoos based on a zoo_id

+++ app/db/schema.sql
@@ -1,4 +1,3 @@
-
 DROP DATABASE IF EXISTS zoomate;
 
 CREATE DATABASE zoomate;
@@ -11,7 +10,7 @@ CREATE TABLE IF NOT EXISTS zoos (
     address VARCHAR(100),
     city VARCHAR(50),
     state VARCHAR(2) ,
-    zip_code TINYINT NOT NULL,
+    zip_code INT NOT NULL,
     phone varchar(20),
     contact_name varchar(255),
     PRIMARY KEY (id)
